Added whitelist command
As some users do not appear in ctx.channel.members, this command serves
as a workaround to permit them to submit article titles.

It is my hypothesis that the users who do not appear in
ctx.channel.members are users who are using mobile devices. However,
I've been (as of yet) unable to confirm that hypothesis.